# api/openapi.yaml
openapi: 3.1.0
info:
  title: API Quiz Bank
  version: 0.1.0
  summary: MVP API contract for governed quiz delivery.
servers:
  - url: http://127.0.0.1:8000
    description: Local MVP API endpoint.
  - url: https://api.valerchik.de
    description: Protected beta route; edge deployments may also require X-API-Key.
paths:
...
  /v1/admin/dashboard:
    get:
      operationId: getAdminDashboard
      summary: Return MVP admin dashboard counts.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
      responses:
        '200':
          description: Admin dashboard counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardResponse'
        '401':
          description: Missing or invalid admin credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/admin/quiz-items:
    get:
      operationId: listAdminQuizItems
      summary: List admin quiz item projections with controlled filters.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, imported, normalized, needs_review, approved, published, monitored, retired, blocked]
        - name: cefr_level
          in: query
          schema:
            type: string
            enum: [A1, A2, B1, B2, C1, C2]
        - name: theme_id
          in: query
          schema:
            type: string
            pattern: '^T(0[1-9]|1[0-8])$'
        - name: source_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Admin quiz item list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQuizItemListResponse'
        '401':
          description: Missing or invalid admin credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/admin/quiz-items/{item_id}:
    get:
      operationId: getAdminQuizItem
      summary: Read one admin quiz item projection.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin quiz item projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQuizItemProjection'
        '404':
          description: Quiz item not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/admin/audit-log:
    get:
      operationId: listAdminAuditLog
      summary: List recent audit events for admin review.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Recent audit events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          description: Missing or invalid admin credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/admin/consumers:
    get:
      operationId: listAdminConsumers
      summary: List consumers, channels and bots with usage counts.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Admin consumer list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConsumerListResponse'
    post:
      operationId: createAdminConsumer
      summary: Create consumer, channel or bot access as owner.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminConsumerCreateRequest'
      responses:
        '200':
          description: Created consumer access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConsumerProjection'
        '403':
          $ref: '#/components/responses/AdminForbidden'
  /v1/admin/consumers/{consumer_id}/visual-settings:
    get:
      operationId: getAdminConsumerVisualSettings
      summary: Read owner-only visual delivery settings for a consumer.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminConsumerId'
      responses:
        '200':
          description: Consumer visual settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVisualSettingsProjection'
        '403':
          $ref: '#/components/responses/AdminForbidden'
        '404':
          description: Consumer not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    patch:
      operationId: patchAdminConsumerVisualSettings
      summary: Update admin-only visual delivery settings as owner.
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminConsumerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminVisualSettingsPatchRequest'
      responses:
        '200':
          description: Updated visual settings with audit evidence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVisualSettingsUpdateResponse'
        '403':
          $ref: '#/components/responses/AdminForbidden'
        '404':
          description: Consumer not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Visual settings payload is invalid.
          content:
            application/json:
              schema:
                type: object
...
components:
  parameters:
    AdminKeyHeader:
      name: X-QuizBank-Admin-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
    AdminConsumerId:
      name: consumer_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    AdminForbidden:
      description: Admin role cannot perform this action.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, reason_code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        reason_code:
          type: string
    SourceTraceability:
      type: object
      additionalProperties: false
      required: [source_id, source_type, provenance_note]
      properties:
        source_id:
          type: string
        source_type:
          type: string
        provenance_note:
          type: string
...
    AdminQuizItemProjection:
      type: object
      additionalProperties: false
      required:
        - item_id
        - language
        - cefr_level
        - theme_id
        - objective_id
        - pattern_id
        - status
        - prompt
        - stem_text
        - options
        - source_traceability
      properties:
        item_id:
          type: string
        language:
          type: string
        cefr_level:
          type: string
        theme_id:
          type: string
        objective_id:
          type: string
        pattern_id:
          type: string
        status:
          type: string
        prompt:
          type: string
        stem_text:
          type: string
        options:
          type: array
          items:
            type: string
        source_traceability:
          $ref: '#/components/schemas/SourceTraceability'
    AdminQuizItemListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminQuizItemProjection'
    AuditLogEntry:
      type: object
      additionalProperties: false
      required: [audit_id, actor, action, entity_type, entity_id, from_status, to_status, reason, created_at]
      properties:
        audit_id:
          type: string
        actor:
          type: string
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        from_status:
          type: string
        to_status:
          type: string
        reason:
          type: string
        created_at:
          type: string
    AuditLogListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
    AdminDashboardResponse:
      type: object
      additionalProperties: false
      required:
        - corpus_status_counts
        - items_by_cefr_level
        - items_by_theme
        - approved_published_count
        - delivery_log_count
        - audit_log_count
        - visual_metrics
      properties:
        corpus_status_counts:
          type: object
          additionalProperties:
            type: integer
        items_by_cefr_level:
          type: object
          additionalProperties:
            type: integer
        items_by_theme:
          type: object
          additionalProperties:
            type: integer
        approved_published_count:
          type: integer
        delivery_log_count:
          type: integer
        audit_log_count:
          type: integer
        visual_metrics:
          $ref: '#/components/schemas/AdminVisualMetricsSummary'
    AdminVisualMetricsSummary:
      type: object
      additionalProperties: false
      required:
        - settings_by_mode
        - event_counts
        - cache_hit_rate
        - fallback_rate
        - generation_count_by_consumer
        - qa_rejection_count
        - estimated_cost_minor_by_consumer
        - estimated_cost_minor_total
      properties:
        settings_by_mode:
          type: object
          additionalProperties:
            type: integer
        event_counts:
          type: object
          additionalProperties:
            type: integer
        cache_hit_rate:
          type: number
        fallback_rate:
          type: number
        generation_count_by_consumer:
          type: object
          additionalProperties:
            type: integer
        qa_rejection_count:
          type: integer
        estimated_cost_minor_by_consumer:
          type: object
          additionalProperties:
            type: integer
        estimated_cost_minor_total:
          type: integer
...
    AdminConsumerCreateRequest:
      type: object
      additionalProperties: false
      required:
        - consumer_id
        - display_name
        - consumer_kind
        - daily_quota_limit
        - allowed_cefr_levels
        - allowed_theme_ids
        - api_key
        - reason
      properties:
        consumer_id:
          type: string
        display_name:
          type: string
        consumer_kind:
          type: string
          enum: [api_client, telegram_channel, telegram_bot, teacher, school]
        daily_quota_limit:
          type: integer
          minimum: 0
        allowed_cefr_levels:
          type: array
          items:
            type: string
            enum: [A1, A2, B1, B2, C1, C2]
        allowed_theme_ids:
          type: array
          items:
            type: string
            pattern: '^T(0[1-9]|1[0-8])$'
        api_key:
          type: string
          minLength: 8
        reason:
          type: string
          minLength: 3
    AdminConsumerProjection:
      type: object
      additionalProperties: false
      required:
        - consumer_id
        - display_name
        - consumer_kind
        - status
        - daily_quota_limit
        - allowed_cefr_levels
        - allowed_theme_ids
        - delivery_count
        - today_quota_used
        - last_delivery_at
      properties:
        consumer_id:
          type: string
        display_name:
          type: string
        consumer_kind:
          type: string
        status:
          type: string
        daily_quota_limit:
          type: integer
        allowed_cefr_levels:
          type: array
          items:
            type: string
        allowed_theme_ids:
          type: array
          items:
            type: string
        delivery_count:
          type: integer
        today_quota_used:
          type: integer
        last_delivery_at:
          type:
            - string
            - 'null'
    VisualDeliveryMode:
      type: string
      enum: [text_only, image_standard, image_branded]
    VisualFallbackPolicy:
      type: string
      enum: [text_only, cache_only, block_visual_delivery]
    AdminVisualSettingsProjection:
      type: object
      additionalProperties: false
      required:
        - consumer_id
        - delivery_mode
        - visual_style
        - branding_preset
        - fallback_policy
        - daily_visual_delivery_limit
        - daily_generation_limit
        - monthly_generation_limit
        - is_active
      properties:
        consumer_id:
          type: string
        delivery_mode:
          $ref: '#/components/schemas/VisualDeliveryMode'
        visual_style:
          type: string
        branding_preset:
          type: string
        fallback_policy:
          $ref: '#/components/schemas/VisualFallbackPolicy'
        daily_visual_delivery_limit:
          type: integer
          minimum: 0
        daily_generation_limit:
          type: integer
          minimum: 0
        monthly_generation_limit:
          type: integer
          minimum: 0
        is_active:
          type: boolean
    AdminVisualSettingsPatchRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        delivery_mode:
          $ref: '#/components/schemas/VisualDeliveryMode'
        visual_style:
          type: string
          minLength: 1
          maxLength: 120
        branding_preset:
          type: string
          minLength: 1
          maxLength: 120
        fallback_policy:
          $ref: '#/components/schemas/VisualFallbackPolicy'
        daily_visual_delivery_limit:
          type: integer
          minimum: 0
          maximum: 10000
        daily_generation_limit:
          type: integer
          minimum: 0
          maximum: 10000
        monthly_generation_limit:
          type: integer
          minimum: 0
          maximum: 100000
        is_active:
          type: boolean
        reason:
          type: string
          minLength: 3
          maxLength: 500
    AdminVisualSettingsUpdateResponse:
      type: object
      additionalProperties: false
      required: [settings, audit]
      properties:
        settings:
          $ref: '#/components/schemas/AdminVisualSettingsProjection'
        audit:
          anyOf:
            - $ref: '#/components/schemas/AuditLogEntry'
            - type: 'null'
    AdminConsumerListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminConsumerProjection'
...